Create LICENSE
I personally hold the copyright on all this code, and I wan't it to be available for everyone to use, but not make it impossible to a future maintainer to fix errors or perform other modifications.  If I got the LGPL right, it should not prevent users of this library from being closed source, and that is my intention.